Move UiLocalesMiddleware to oidc_provider app
